requirements.txt

# Core ML libraries
pandas==2.3.3
numpy==2.3.3
scikit-learn==1.7.2
xgboost>=2.0.0,<3.0.0

# Environment for experimentation
ipykernel==6.30.1

# MLflow for experiment tracking
mlflow==3.6.0

# DVC for data versioning
dvc==3.63.0
dvc-gdrive  # if using Google Drive remote

# Visualization
matplotlib==3.10.6

# Configuration management
pyyaml==6.0.3

# Optional: Additional ML libraries
# xgboost==3.1.1
# lightgbm==4.6.0
# catboost==1.2.8
# shap==0.49.1  # for model interpretability